+++ src/main/java/com/owlmaddie/utils/VillagerEntityAccessor.java
@@ -1,0 +1,7 @@
+package com.owlmaddie.utils;
+
+import net.minecraft.village.VillagerGossips;
+
+public interface VillagerEntityAccessor {
+    VillagerGossips getGossip();
+}